VxMark HWTA: Wire up additional device trackers
